Change lightmapper to only use one image

+++ src/common/rendering/vulkan/textures/vk_texture.cpp
@@ -219,9 +219,9 @@ void VkTextureManager::CreateLightmap(int newLMTextureSize, int newLMTextureCoun
 
 	auto cmdbuffer = fb->GetCommands()->GetTransferCommands();
 
-	if (newPixelData.Size() >= w * h * count * 3)
+	if (newPixelData.Size() >= (size_t)w * h * count * 3)
 	{
-		assert(newPixelData.Size() == w * h * count * 3);
+		assert(newPixelData.Size() == (size_t)w * h * count * 3);
 
 		int totalSize = w * h * count * pixelsize;
